% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessAtoms.R
\name{access_atoms}
\alias{access_atoms}
\title{Annotate surface atoms}
\usage{
access_atoms(
  pdb,
  chains = NULL,
  minacc = get.pepit("ACC"),
  probeRadius = get.pepit("radius"),
  add = get.pepit("ADD"),
  select = FALSE
)
}
\arguments{
\item{pdb}{a pdb structure returned by bio3d::read.pdb}

\item{chains}{vector of chains to be processed}

\item{minacc}{accessibility threshold. b field set to 0 if SASA is less than ACC.}

\item{probeRadius}{probe radius in Angstroms for the FreeSASA application}

\item{add}{type of carbon atoms to be added}

\item{select}{boolean for keeping atoms in output pdb}
}
\value{
modified pdb structure (b field with SASA values)
}
\description{
It replaces the b field by the solvent accessible surface area

protein chains are processed independantly

if chains!=NULL, only b field of chains in chains are set to acc.

b field of the other chains are set to 0

it implies that binding sites between chains in contact are accessible
}
