% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildGraph_ho}
\alias{buildGraph_ho}
\title{Constructs correspondence graph}
\usage{
buildGraph_ho(X, XResno, Y, YResno, V, deltadist, mindist, maxdist, maxgap)
}
\arguments{
\item{X}{matrix Nx3}

\item{Y}{matrix Nx3}

\item{V}{matrix Nvx2}

\item{deltadist}{double}

\item{mindist}{double}

\item{maxdist}{double}
}
\description{
Constructs correspondence graph
}
