% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildGraph_ho2}
\alias{buildGraph_ho2}
\title{Constructs correspondence graph}
\usage{
buildGraph_ho2(X, XResno, XResInd, Y, YResno, YResInd, V, deltadist, maxgap)
}
\arguments{
\item{X}{matrix Nx3}

\item{Y}{matrix Nx3}

\item{V}{matrix Nvx2}

\item{deltadist}{double}

\item{mindist}{double}

\item{maxdist}{double}
}
\description{
Constructs correspondence graph
}
