% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCliques.R
\name{cliques}
\alias{cliques}
\title{cliques}
\usage{
cliques(
  X,
  XProp,
  Y,
  YProp,
  deltadist,
  mindist,
  maxdist,
  types,
  verbose = FALSE
)
}
\arguments{
\item{X}{atom coordinate matrix from target (Nx3)}

\item{XProp}{atom label vector from target (N)}

\item{Y}{atom coordinate matrix from query (a binding site) (Mx3)}

\item{YProp}{atom label vector from query  (M)}

\item{deltadist}{precision parameter}

\item{mindist}{minimum atom distance considered}

\item{maxdist}{maximal atom distance considered}

\item{types}{atom types from which to build graph}

\item{precision}{parameter ?}
}
\value{
list of clusters : a cluster is a list of X indices (cluster$I) and of Y indices (cluster$J)
}
\description{
cliques
}
