% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCliques.R
\name{cliques_ho2}
\alias{cliques_ho2}
\title{cliques for hot spots application without based on complete residue (and not atom separatly)}
\usage{
cliques_ho2(
  X,
  XResname,
  XResno,
  XResInd,
  Y,
  YResname,
  YResno,
  YResInd,
  deltadist,
  types,
  verbose = FALSE
)
}
\arguments{
\item{X}{atom coordinate matrix from target (Nx3)}

\item{XResname}{residue name vector from target (N)}

\item{XResno}{residue number vector from target (N)}

\item{Y}{atom coordinate matrix from query (a binding site) (Mx3)}

\item{YResname}{residue name vector from query (N)}

\item{YResno}{residue number vector from query (N)}

\item{deltadist}{precision parameter}

\item{types}{atom types from which to build graph}

\item{XProp}{atom label vector from target (N)}

\item{YProp}{atom label vector from query  (M)}

\item{precision}{parameter ?}

\item{mindist}{minimum atom distance considered}

\item{maxdist}{maximal atom distance considered}
}
\value{
list of clusters : a cluster is a list of X indices (cluster$I) and of Y indices (cluster$J)
}
\description{
cliques for hot spots application without based on complete residue (and not atom separatly)
}
