% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\title{Atom typing}
\usage{
encode(pdb, cutoff = get.pepit("HSECUTOFF"), pc.code = get.pepit("PCCODE"))
}
\arguments{
\item{pdb}{a pdb structure returned by bio3d::read.pdb}
}
\value{
a data frame
}
\description{
The function produces a data frame with pdb ATOM information

and add columns with HSE values, physico-chemcal encoding of residues and an atom type column:

alpha carbons changed to A, beta carbons changed to B,

other backbone carbons remains C,

aromatic carbons changed to a, other side-chain atoms changed to

to lower case (o,n,c)

ligand chain elements are not modified
}
