% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendCliques.R
\name{extend_cliques}
\alias{extend_cliques}
\title{extends cliques}
\usage{
extend_cliques(X, XProp, Y, YProp, cliques, deltadist, score_function)
}
\arguments{
\item{X}{atom coordinate matrix from target (Nx3)}

\item{XProp}{atom label vector from target (N)}

\item{Y}{atom coordinate matrix from query (a binding site) (Mx3)}

\item{YProp}{atom label vector from query  (M)}

\item{cliques}{list of mappings that are lists of corresponding atom indices ($I and $J)}

\item{deltadist}{precision parameter}
}
\value{
list of clusters : a cluster is a list of X indices (cluster$I) and of Y indices (cluster$J)
}
\description{
extends cliques
}
