% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendCliques.R
\name{extend_cliques_ho}
\alias{extend_cliques_ho}
\title{extends cliques for hot spots application}
\usage{
extend_cliques_ho(
  X,
  XProp,
  XResname,
  XResno,
  Y,
  YProp,
  YResname,
  YResno,
  cliques,
  deltadist,
  score_function = get.pepit("SCORE"),
  verbose = FALSE
)
}
\arguments{
\item{X}{atom coordinate matrix from target (Nx3)}

\item{XProp}{atom label vector from target (N)}

\item{XResname}{residue name vector from target (N)}

\item{XResno}{residue number vector from target (N)}

\item{Y}{atom coordinate matrix from query (a binding site) (Mx3)}

\item{YProp}{atom label vector from query  (M)}

\item{YResname}{residue name vector from query (N)}

\item{YResno}{residue number vector from query (N)}

\item{cliques}{list of mappings that are lists of corresponding edge indices}

\item{deltadist}{precision parameter}
}
\value{
list of clusters : a cluster is a list of X indices (cluster$I) and of Y indices (cluster$J)
}
\description{
extends cliques for hot spots application
}
