% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBindingSite.R
\name{get_binding_sites}
\alias{get_binding_sites}
\title{create pdb file with binding site atoms marked with b field>0 + ligand}
\usage{
get_binding_sites(
  pdb,
  target_chains,
  ligand_chains,
  cutoff = 5,
  add = "calpha",
  add.ligand = TRUE,
  surface = -1
)
}
\arguments{
\item{pdb}{pdb file as read by bio3d}

\item{target_chains}{vector of selected chains as target}

\item{ligand_chains}{vector of selected chains as ligand}

\item{cutoff}{distance cutoff for defining contact atoms}

\item{add}{type of carbon atoms to be added to the binding sites}

\item{add.ligand}{add ligand to the output pdb}

\item{surface}{if >0 keep only surface atoms on returned pdb}
}
\value{
pdb bio3d structure
}
\description{
create pdb file with binding site atoms marked with b field>0 + ligand
}
