% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBindingSite.R
\name{get_binding_sites_2}
\alias{get_binding_sites_2}
\title{create pdb file with binding site atoms marked with b field>0 + ligand}
\usage{
get_binding_sites_2(
  pdb,
  target_chains,
  ligand_chains,
  cutoff = get.pepit("CONTACT"),
  add = get.pepit("ADD")
)
}
\arguments{
\item{pdb}{pdb file as read by bio3d}

\item{target_chains}{vector of selected chains as target}

\item{ligand_chains}{vector of selected chains as ligand}

\item{cutoff}{distance cutoff for defining contact atoms}

\item{add}{type of atoms to be added to the binding sites}
}
\value{
pdb bio3d structure
}
\description{
create pdb file with binding site atoms marked with b field>0 + ligand
}
