% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCarbons.R
\name{get_carbons}
\alias{get_carbons}
\title{Annotate surface atoms}
\usage{
get_carbons(pdb, inds, add = "calpha")
}
\arguments{
\item{pdb}{a pdb structure returned by bio3d::read.pdb}

\item{inds}{a vector of atom indices (of pdb$atom) to be processed}

\item{add}{"calpha", "cab" or "" add alpha carbons of residues with accessible atoms}
}
\value{
atom indices , list of class select
}
\description{
it returns indices of alpha carbons of residues of given atoms, (or alpha and cbeta atoms)

protein chains are processed independantly
}
