% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCliques.R
\name{get_mapping}
\alias{get_mapping}
\title{convert list of atom pairs (edges of bipartite mapping) to a list of atom pairs}
\usage{
get_mapping(clusters, N)
}
\arguments{
\item{clusters}{: vector list of atom pairs}
}
\value{
list of mappings : a mapping is a list of X indices (mapping$I) and of corresponding Y indices (mapping$J)
}
\description{
convert list of atom pairs (edges of bipartite mapping) to a list of atom pairs
}
