% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{p_values}
\alias{p_values}
\title{evaluate pvalues from all scores}
\usage{
p_values(score)
}
\arguments{
\item{score}{vector of scores}
}
\value{
vector of p values
}
\description{
evaluate pvalues from all scores
}
