% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPDB.R
\name{read_pdb}
\alias{read_pdb}
\title{read a PDB file into a list}
\usage{
read_pdb(pdbfile, chain = "", access = FALSE, resi = integer(0))
}
\arguments{
\item{pdbfile}{pdb file}

\item{chain}{vector of selected chains}

\item{access}{boolean to select atoms with b-factor > 0}

\item{resi}{vector of residu numbers of selected residus}
}
\value{
list of atoms (PepiT format)
}
\description{
read a PDB file into a list
}
