% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCliques.R, R/scores.R
\name{rmsd}
\alias{rmsd}
\title{Compute rmsd between X and Y}
\usage{
rmsd(X, Y)

rmsd(X, Y)
}
\arguments{
\item{X}{matrix Nx3}

\item{Y}{matrix Nx3}
}
\description{
Compute rmsd between X and Y

Compute rmsd between X and Y
}
