% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{score_clusters}
\alias{score_clusters}
\title{evaluate cliques and clusters}
\usage{
score_clusters(
  clusters,
  X,
  Y,
  score_function = get.pepit("SCORE"),
  deltadist = 1
)
}
\arguments{
\item{clusters}{list of mappings : a mapping is a set of correspondence identifiers}

\item{X}{atom coordinate matrix from target (Nx3)}

\item{Y}{atom coordinate matrix from query (a binding site) (Mx3)}

\item{score_function}{score optimized by maximal matching procedure}
}
\value{
list of mapping criteria :
}
\description{
evaluate cliques and clusters
}
