% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessAtoms.R
\name{surface_atoms}
\alias{surface_atoms}
\title{select surface atoms}
\usage{
surface_atoms(pdb, chains = NULL, minacc = ACC)
}
\arguments{
\item{pdb}{a pdb structure returned by bio3d::read.pdb}

\item{chains}{vector of chains to be processed}

\item{minacc}{double accessibility threshold}
}
\value{
pdb structure
}
\description{
it selects accessible atoms (access>=minacc) and return a pdb structure
}
