% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessAtoms.R
\name{trim_non_acc_atoms}
\alias{trim_non_acc_atoms}
\title{select surface atoms}
\usage{
trim_non_acc_atoms(pdb, chains = NULL, minacc = get.pepit("ACC"))
}
\arguments{
\item{pdb}{a pdb structure returned by bio3d::read.pdb}

\item{chains}{vector of chains to be processed}

\item{minacc}{double accessibility threshold}
}
\value{
pdb structure with non accessible atoms removed
}
\description{
it selects accessible atoms (access>=minacc) and return a pdb structure
}
