% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeAtoms.R
\name{type_atoms}
\alias{type_atoms}
\title{Atom typing}
\usage{
type_atoms(pdb)
}
\arguments{
\item{pdb}{a pdb structure returned by bio3d::read.pdb}
}
\value{
modified pdb structure
}
\description{
The function modifies the element symbol field (elesy)

alpha carbons changed to a, beta carbons changed to b

aromatic carbons changed to A, other carbons remain C
}
