% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vertex}
\alias{vertex}
\title{Constructs vertices of correspondence graph}
\usage{
vertex(XProp, YProp, mode, size, hse)
}
\arguments{
\item{XProp}{data frame  N nrows}

\item{YProp}{data frame  M nrows}

\item{mode}{0 atom types comparison, 1 all features' comparison}

\item{size}{estimated initial size of V}

\item{V}{output matrix Nvx2 of vertices}
}
\description{
data.frame(eleno, elety, resid, chain, resno, insert, x, y, z, pc, hseu, hsed)
}
