% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vertex_ho}
\alias{vertex_ho}
\title{Constructs vertices of correspondence graph}
\usage{
vertex_ho(XProp, XResname, YProp, YResname)
}
\arguments{
\item{XProp}{size N vector of atom types}

\item{XResname}{size N vector of residue names}

\item{YProp}{size M vector of atom types}

\item{YResname}{size M vector of residue names}

\item{V}{output, matrix Nvx2 of vertices}
}
\description{
Constructs vertices of correspondence graph
}
