% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vertex_ho2}
\alias{vertex_ho2}
\title{Constructs vertices of correspondence graph based on residues}
\usage{
vertex_ho2(XResno, XResname, YResno, YResname)
}
\arguments{
\item{XResno}{size N vector of residue number}

\item{XResname}{size N vector of residue names}

\item{YResno}{size M vector of residue number}

\item{YResname}{size M vector of residue names}

\item{V}{output, matrix Nvx2 of vertices}
}
\description{
Constructs vertices of correspondence graph based on residues
}
