% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessAtoms.R
\name{access_atoms}
\alias{access_atoms}
\title{Annotate surface atoms}
\usage{
access_atoms(
  pdb,
  chains = NULL,
  minacc = ACC,
  probeRadius = radius,
  surface = FALSE,
  add = "calpha"
)
}
\arguments{
\item{pdb}{a pdb structure returned by bio3d::read.pdb}

\item{chains}{vector of chains to be processed}

\item{minacc}{accessibility threshold. b field set to 0 if SASA is less than ACC.}

\item{probeRadius}{probe radius in Angstroms for the FreeSASA application}

\item{surface}{boolean for keeping atoms in output pdb}

\item{add}{include carbon alpha of residues with accessible atoms}
}
\value{
modified pdb structure (b field with SASA values)
}
\description{
It replaces the b field by the solvent accessible surface area

protein chains are processed independantly

it implies that binding sites between chains in contact are accessible
}
