% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superposeSites.R
\name{superpose_sites}
\alias{superpose_sites}
\title{superpose query (binding site) onto the target. 
The ligand is kept while the binding site is removed.}
\usage{
superpose_sites(cluster, query, target, pepchain)
}
\arguments{
\item{cluster}{indices of mapping edges}

\item{query}{pdb structure containing bs and ligand as returned by bio3d}

\item{target}{pdb structure of target protein as returned by bio3d}
}
\value{
a pdb structure containing the moved ligand
}
\description{
superpose query (binding site) onto the target. 
The ligand is kept while the binding site is removed.
}
