% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superposeSites.R
\name{superpose_sites}
\alias{superpose_sites}
\title{superpose query (binding site) onto the target. 
The ligand is kept while the binding site is removed.}
\usage{
superpose_sites(cluster, bs, target, peptide)
}
\arguments{
\item{cluster}{indices of mapping edges}

\item{bs}{data frame containing bs features (mobile)}

\item{target}{data frame containing target features (fixed)}

\item{peptide}{pdb file containing peptide (or ligand)}
}
\value{
a pdb structure containing the moved ligand
}
\description{
superpose query (binding site) onto the target. 
The ligand is kept while the binding site is removed.
}
\examples{
             ligand.moved = superpose_sites2(clusters[[i]], bs.data, target.data, peptide)
}
