% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vertex}
\alias{vertex}
\title{Constructs vertices of correspondence graph}
\usage{
vertex(XProp, YProp)
}
\arguments{
\item{XProp}{vecteur of atom types N}

\item{YProp}{vecteur of atom types N}

\item{V}{matrix Nvx2 of vertices}
}
\description{
Constructs vertices of correspondence graph
}
